+++ Runtime/Components/ImageComponent.cs
@@ -9,50 +9,26 @@ namespace UCGUI
 {
     /// <summary>
     /// UCGUI's Image Component.
-    /// Noteworthy formatting functions:
-    /// <list type="bullet">
-    /// <item><description><see cref="Sprite(UnityEngine.Sprite,UnityEngine.UI.Image.Type?,float)"/> - Sets the image's sprite with optional <see cref="Image.Type"/> and <see cref="Image.pixelsPerUnitMultiplier"/>.</description></item>
-    /// <item><description><see cref="GraphicComponent{T}.Color(UnityEngine.Color,bool)"/> - Sets the color. Allows you to also keep the previous alpha value if adding 'true'.</description></item>
-    /// <item><description><see cref="GraphicComponent{T}.Alpha"/> - Sets the alpha.</description></item>
-    /// <item><description><see cref="Clear"/> - Sets the sprite to null.</description></item>
-    /// <item><description><see cref="NativeSize()"/> - Sets the RectTransform's size to the pixel size of the sprite. Use <see cref="NativeSize(Vector2)"/> or <see cref="NativeSize(float, float)"/> to additionally scale the image by a factor.</description></item>
-    /// <item><description><see cref="ImageType"/> - <see cref="Image.Type"/> used to render the image.</description></item>.
-    /// <item><description><see cref="Filled"/> - Allows you to specify a fill method, as well an optional fill amount. (Defaults to '1f').</description></item>.
-    /// <item><description><see cref="FillAmount"/> - Adjusts the fill amount.</description></item>.
-    /// <item><description><see cref="PixelsPerUnitMultiplier"/> - Adjusts the <see cref="Image.pixelsPerUnitMultiplier"/> of the image. For example when using <see cref="Image.Type.Sliced"/>.</description></item>.
-    /// <item><description><see cref="GraphicComponent{T}.RaycastTarget"/> - Determines whether the image acts a raycast target.</description></item>.
-    /// <item><description><see cref="GraphicComponent{T}.ToggleVisibility"/> - Enables / disables the image render.</description></item>.
-    /// <item><description><see cref="AddAnimator"/> - Adds a <see cref="SpriteAnimator"/> behaviour.</description></item>.
-    /// <item><description><see cref="Material"/> - Sets the material.</description></item>.
-    /// </list>
     /// <para>
-    /// Also implements <see cref="ICopyable{T}"/> which allows <see cref="ICopyable{T}.CopyFrom"/> and <see cref="ICopyable{T}.Copy"/>.
+    /// Implements <see cref="ICopyable{T}"/> which allows <see cref="ICopyable{T}.CopyFrom"/> and <see cref="ICopyable{T}.Copy"/>.
     /// </para>
     /// </summary>
-    public partial class ImageComponent : GraphicComponent<ImageComponent>, ICopyable<ImageComponent>, IStylable<ImageComponent, ImageStyle>
+    public class ImageComponent : GraphicComponent<ImageComponent>, ICopyable<ImageComponent>, IStylable<ImageComponent, ImageStyle>
     {
         protected ImageComponent() {}
         
         private Image _image;
         protected static readonly string NamePrefix = "Image";
         
-        private InputAction _toggleInputAction;
-        
         protected SpriteAnimator animator;
 
-        public override void Awake()
+        protected override void Awake()
         {
             base.Awake();
             DisplayName = NamePrefix;
             
             _image = gameObject.GetOrAddComponent<Image>();
             this.Size(Defaults.Image.DefaultSize);
-        }
-
-        public virtual void Start()
-        {
-            if (_toggleInputAction != null)
-                _toggleInputAction.performed += _ => ToggleVisibility();
         }
 
         /// <summary>
@@ -136,21 +112,18 @@ namespace UCGUI
             _image.pixelsPerUnitMultiplier = multiplier;
             return this;
         }
-
-        public Image GetImage()
+        
+        public ImageComponent Maskable(bool maskable)
         {
-            return _image;
+            _image.maskable = maskable;
+            return this;
         }
 
-        public override Graphic GetGraphic()
-        {
-            return _image;
-        }
+        public Image GetImage() { return _image; }
 
-        public bool HasImage()
-        {
-            return _image.sprite;
-        }
+        public override Graphic GetGraphic() { return _image; }
+
+        public bool HasImage() { return _image.sprite; }
 
         public new ImageComponent Copy(bool fullyCopyRect = true)
         {
@@ -162,7 +135,6 @@ namespace UCGUI
         {
             base.CopyFrom(other, fullyCopyRect);
             CopyImageProperties(other.GetImage(), this);
-            CopyMaterial(other, this);
             CopyAnimator(other, this);
             return this;
         }
@@ -174,15 +146,6 @@ namespace UCGUI
             copyImage.PixelsPerUnitMultiplier(image.pixelsPerUnitMultiplier);
         }
 
-        public static void CopyMaterial(ImageComponent other, ImageComponent copyImage)
-        {
-            var defaultMat = Graphic.defaultGraphicMaterial;
-            if (other.GetImage().material && other.GetImage().material != defaultMat)
-            {
-                copyImage.Material(other.GetImage().material);
-            }
-        }
-
         public static void CopyAnimator(ImageComponent other, ImageComponent copyImage)
         {
             if (other.animator)
@@ -190,30 +153,6 @@ namespace UCGUI
                 copyImage.AddAnimator();
                 copyImage.animator.CopyFrom(other.animator);
             }
-        }
-
-        public ImageComponent ToggleVisibilityUsing(InputAction action)
-        {
-            _toggleInputAction = action;
-            _toggleInputAction?.Enable();
-            return this;
-        }
-
-        private void OnEnable()
-        {
-            _toggleInputAction?.Enable();
-        }
-
-        private void OnDisable()
-        {
-            _toggleInputAction?.Disable();
-        }
-
-        // TODO: For more flexible and efficient use, allowing controlling of instances, we might want to use MaterialPropertyBlock in the future
-        public ImageComponent Material(Material material)
-        {
-            _image.material = material;
-            return this;
         }
 
         /// <summary>